feat(api): Add display_name and seller_rating to seller accounts
- Migration 0005: Add display_name (VARCHAR(200)) and seller_rating (DECIMAL(3,2)) to seller_accounts
- Update SellerAccountRow to include display_name and seller_rating fields
- Update all seller_accounts queries (get_by_owner_id, update_trust_level, etc.) to SELECT new fields
- Update row_to_summary() to use display_name and seller_rating from JOIN
- Update get_listing() and fetch_rows() to LEFT JOIN seller_accounts
- Update update_listing_status() to RETURN marketplace fields
- Change seller_rating type in API contract from f32 to f64 (matches DB precision)
- All tests pass ✅

+++ backend/crates/api-contract/src/listing.rs
@@ -121,7 +121,7 @@ pub struct ListingSummary {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub seller_name: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub seller_rating: Option<f32>,
+    pub seller_rating: Option<f64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub seller_verified: Option<bool>,
 }
